Organize exporter formatters under their format namespace
