+++ app/models/projects/exports/formatters/pdf/required_disk_space.rb
@@ -27,15 +27,21 @@
 #
 # See COPYRIGHT and LICENSE files for more details.
 #++
-module WorkPackage::Exports
+module Projects::Exports
   module Formatters
-    class Hours < ::Exports::Formatters::Default
-      def self.apply?(name, export_format)
-        name.to_sym == :spent_hours && export_format == :pdf
-      end
+    module PDF
+      class RequiredDiskSpace < ::Exports::Formatters::Default
+        def self.apply?(attribute, export_format)
+          export_format == :pdf && attribute.to_sym == :required_disk_space
+        end
 
-      def format_value(value, _options)
-        DurationConverter.output(value)
+        ##
+        # Takes a project and returns the formatted value if the required disk space is greater than 0.
+        def format(project, **)
+          return "" unless project.required_disk_space.to_i > 0
+
+          number_to_human_size(project.required_disk_space, precision: 2)
+        end
       end
     end
   end
